+++ miracle/__init__.py
@@ -1,1 +1,1 @@
-VERSION = '1.0.9.dev1'
+VERSION = '1.0.9'